# medical-appointments/openapi.yml
openapi: 3.0.0
components:
  schemas:
    CreateAppointmentRequest:
      type: object
      required:
        - insuredId
        - scheduleId
        - countryISO
      properties:
        insuredId:
          type: string
          example: "01234"
        scheduleId:
          type: number
          example: 100
        countryISO:
          type: string
          enum:
            - PE
            - CL
          example: PE
    AppointmentCreatedResponse:
      type: object
      properties:
        message:
          type: string
          example: Appointment scheduled. Processing...
        id:
          type: string
          example: "123"
    AppointmentList:
      type: array
      items:
        type: object
        properties:
          id:
            type: string
          insuredId:
            type: string
          scheduleId:
            type: number
          countryISO:
            type: string
          status:
            type: string
          createdAt:
            type: string
            format: date-time
          updatedAt:
            type: string
            format: date-time
    ErrorResponse400Create:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error
          example: Missing required fields
        details:
          type: string
          description: Detalles adicionales del error
          example: "The fields insuredId, scheduleId and countryISO are required"
    ErrorResponse422Country:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error
          example: "Error countryISO "
        details:
          type: string
          description: Detalles adicionales del error
          example: countryISO must be PE or CL
    ErrorResponse400GetIn:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error
          example: Missing required fields
        details:
          type: string
          description: Detalles adicionales del error
          example: Missing insuredId param
    ErrorResponse500:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error
          example: Internal Server Error
        details:
          type: string
          description: Detalles adicionales del error
          example: Internal Server Error
  securitySchemes: {}
info:
  title: API de Citas Médicas
  description: API REST para registrar y consultar citas médicas.
  version: 1.0.0
paths:
  /appointments:
    post:
      operationId: appointment
      summary: Crear una cita médica
      description: Registra una cita en DynamoDB y publica un evento SNS.
      requestBody:
        description: Datos del agendamiento
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "202":
          description: Cita registrada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentCreatedResponse"
        "400":
          description: Error en los datos de entrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse400Create"
        "422":
          description: Error en los datos de entrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse422Country"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse500"
  /appointments/get:
    get:
      operationId: appointmentGet
      summary: Consultar citas de un asegurado
      description: Devuelve todas las citas registradas de un asegurado dado.
      parameters:
        - name: insuredId
          in: query
          required: true
          description: ID del asegurado
          schema:
            type: string
            example: "01234"
      responses:
        "200":
          description: Lista de citas encontradas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
        "400":
          description: Parámetro insuredId faltante
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse400GetIn"
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse500"